pyproject.toml
[project]
name = "amazon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "celery>=5.3.0",
]
